\name{derivPsiBiSquare}
\alias{derivPsiBiSquare}
\title{Derivative of Bisquare Psi Function}
\description{
Calculates residual from the first derivative of Bisquare psi function}
\usage{
derivPsiBiSquare(r,c)}
\details{
Calculates residual from the first derivative of Bisquare psi function}
\arguments{
\item{r}{Value to be evaluated by function}
\item{c}{Tuning constant}
}
\value{
 r:	Value to be evaluated by function

 c:	Tuning constant
}                                                                                                                                             \author{Ian M. Johnson \email{ian@alpha-analysis.com}}
\note{\url{http://www.alpha-analysis.com/robustreg.html}}
\references{Tukey}
\seealso{psiBiSquare()}

\keyword{regression}
