\name{derivPsiHuber}
\alias{derivPsiHuber}
\title{Derivative of Huber Psi Function}
\description{
Calculates residual from the first derivative of Huber psi function}
\usage{
derivPsiHuber(r,c)}
\details{
Calculates residual from the first derivative of Huber psi function}
\arguments{
\item{r}{Value to be evaluated by function}
\item{c}{Tuning constant}
}
\value{
 r:	Value to be evaluated by function

 c:	Tuning constant
}                                                                                                                                             \author{Ian M. Johnson \email{ian@alpha-analysis.com}}
\note{\url{http://www.alpha-analysis.com/robustreg.html}}
\references{Huber}
\seealso{psiHuber()}

\keyword{regression}

