\name{mad_rcpp}
\alias{mad_rcpp}
\docType{package}
\title{
Median Absolute Deviation (MAD)
}
\description{
Rcpp fast implementation of median absolute deviation (MAD)
}
\usage{
mad_rcpp(r,scale_factor = 1.4826)
}
\arguments{
\item{r}{A numeric vector}
\item{scale_factor}{Scale factor}
}
\examples{
\dontrun{
mad_rcpp(r,scale_factor = 1.4826)
}
}
