\name{median_rcpp}
\alias{median_rcpp}
\docType{package}
\title{
Median
}
\description{
Rcpp fast implementation of median
}
\usage{
median_rcpp(x)
}
\arguments{
\item{x}{A numeric vector containing the values whose median is to be computed.}
}
\author{Ian M. Johnson}
\examples{
median_rcpp(1:100)
}
