\name{psiBS_rcpp}
\alias{psiBS_rcpp}
\docType{package}
\title{
Tukey's Bisquare Psi Function
}
\description{
Rcpp fast implementation of Tukey's Bisquare psi function
}
\usage{
psiBS_rcpp(r,c)
}
\arguments{
\item{r}{A numeric vector}
\item{c}{Tuning constant}
}
\author{Ian M. Johnson}
\examples{
\dontrun{
psiBS_rcpp(r,c)
}
}
