\name{psiHuber_rcpp}
\alias{psiHuber_rcpp}
\docType{package}
\title{
Huber Psi Function
}
\description{
Rcpp fast implementation of Huber's Psi Function
}
\usage{
psiHuber_rcpp(r,c)
}
\arguments{
\item{r}{A numeric vector}
\item{c}{Tuning constant}
}
\author{Ian M. Johnson}
\examples{
\dontrun{
psiHuber_rcpp(r,c)
}
}
